finished running regression v1, starting to analyze results

+++ scripts/linreg_single_policy.py
@@ -2,8 +2,12 @@
 Generate datasets for notebook 5 - first version of regression
 """
 
-from covid_project.regression_funcs_bins import generate_dataset_group_single_policy
+from covid_project.regression_funcs_bins import generate_dataset_group_single_policy, fit_ols_model_single_policy, get_single_policy_regression_data
+from covid_project.data_utils import get_all_policies
 from covid_project.policy_mappings import policy_dict_v2
+import os
+from tqdm.auto import tqdm
+import json
 import argparse
 
 all_bins = [
@@ -22,32 +26,40 @@ def run_model_on_policies(bins,
     
     results = dict()
     for policy in tqdm(all_policies, desc='running models'):
-        suc, data = get_processed_data(policy, bins)
+        suc, data = get_single_policy_regression_data(policy, bins)
         if not suc:
             print(f"[ERROR] data read failed: bins={bins}, policy={policy}, var={dep_var}")
             continue
-        res = fit_ols_model_single_policy(data,
-                                          policy,
-                                          dep_var,
-                                          True)
+        try:
+            res = fit_ols_model_single_policy(data,
+                                              policy,
+                                              dep_var,
+                                              True)
+        except Exception as e:
+            print(f"ERROR: {e}")
+            print(f"policy: {policy}; bins: {bins}")
         results[policy] = res
     return results
 
 def run_batch_of_models(all_bins,
                         all_policies,
                         dep_vars,
-                        save_path="./data/regression_results/"):
+                        save_path="./data/regression_results_single_policy_bins/",
+                        force_run = False):
     if not os.path.exists(save_path):
         os.makedirs(save_path)
         
-    for bins_list in tqdm(all_bins, desc="looping through bins"):
-        for var in tqdm(dep_vars, desc="looping through dependent variables"):
+    for bins_list in all_bins:
+        for var in dep_vars:
+            print(f"running models for {var} with bins {bins_list}")
+            filename = var + "_bins=" + ''.join([str(b[0])+"-"+str(b[1])+"_" for b in bins_list])[:-1] + ".json"
+            full_path = save_path + filename
+            if os.path.exists(full_path) and not force_run:
+                continue
             results = run_model_on_policies(bins=bins_list,
                                             all_policies=all_policies,
                                             dep_var=var,
                                             pbar=True)
-            filename = var + "_bins=" + ''.join([str(b[0])+"-"+str(b[1])+"_" for b in bins_list])[:-1] + ".json"
-            full_path = save_path + filename
             
             with open(full_path, "w") as f:
                 json.dump(results, f, indent=2)
@@ -73,6 +85,7 @@ def run_models():
     run_batch_of_models(all_bins=all_bins,
                     all_policies=all_policies,
                     dep_vars=dep_vars,)
+
 if __name__ == '__main__':
     
     parser = argparse.ArgumentParser(description='Generate datasets or run regression models.')
